Estandariza estilos de Perfil y corrige fetch en Estadísticas de Voluntariado
