+++ frontend/src/components/Map/VolunteeringMapModal/VolunteeringMapModal.jsx
@@ -2,7 +2,7 @@ import React, { useState, useMemo } from "react";
 import { useNavigate } from "react-router-dom";
 import "./VolunteeringMapModal.css";
 import ConfirmAlert from "../../alerts/ConfirmAlert";
-import { InscribeIntoVolunteering } from "../../../services/volunteering/VolunteeringService";
+import { InscribeIntoVolunteering } from "../../../services/voluntariado/voluntariadoService";
 import { SuccessAlert, WrongAlert } from "../../../utils/ToastAlerts";
 import { useAuth } from "../../../context/AuthContext";
