+++ frontend/src/components/Profile/ProfileInfo/ProfileInfo.jsx
@@ -1,8 +1,6 @@
 import React, { useEffect, useState } from "react";
 import "./ProfileInfo.css";
-import iconoCiudad from "../../../assets/photos/icono-ciudad.png";
-import iconoTelefono from "../../../assets/photos/icono-telefono2.png";
-import iconoFechaNacimiento from "../../../assets/photos/icono-fecha-nacimiento.png";
+import { FaBuilding, FaCalendar, FaPhoneSquareAlt } from "react-icons/fa";
 
 function ProfileInfo({ user }) {
   const [fecha, setFecha] = useState("");
@@ -28,16 +26,14 @@ function ProfileInfo({ user }) {
   return (
     <div className="profile-info">
       <div className="profile-info-item">
-        <img src={iconoCiudad} alt="Ciudad" className="profile-info-icon" />
+        <FaBuilding className="profile-info-icon" />
         <h4>Ciudad</h4>
         <p>{user?.ciudad || "No especificada"}</p>
       </div>
 
       {user?.rol !== "CREADOR" && (
         <div className="profile-info-item profile-info-item-centered">
-          <img
-            src={iconoFechaNacimiento}
-            alt="Fecha de Nacimiento"
+          <FaCalendar
             className="profile-info-icon"
           />
           <h4>Fecha de Nacimiento</h4>
@@ -46,7 +42,7 @@ function ProfileInfo({ user }) {
       )}
 
       <div className="profile-info-item">
-        <img src={iconoTelefono} alt="Teléfono" className="profile-info-icon" />
+        <FaPhoneSquareAlt className="profile-info-icon" />
         <h4>Teléfono</h4>
         <p>{user?.telefono || "No disponible"}</p>
       </div>
