+++ frontend/src/components/Profile/ProfileStats/ProfileStats.css
@@ -1,22 +1,21 @@
 .profile-stats {
   display: flex;
-  gap: 1.25rem;
+  gap: 0.75rem;
   padding-bottom: 2rem;
-  border-bottom: 1px solid rgba(0, 0, 0, 0.08);
+  border-bottom: 1px solid var(--color-border);
   margin-bottom: 2rem;
   width: 100%;
   flex-wrap: wrap;
 }
 
 .profile-stat {
-  background: linear-gradient(135deg, rgba(0, 122, 255, 0.1) 0%, rgba(0, 122, 255, 0.05) 100%);
+  background: var(--color-primary-light);
   backdrop-filter: blur(10px);
   -webkit-backdrop-filter: blur(10px);
   padding: 1.5rem;
   flex: 1;
   min-width: 150px;
-  border-radius: 16px;
-  border: 1px solid rgba(0, 122, 255, 0.2);
+  border-radius: var(--radius-medium);
   text-align: center;
   transition: all 0.3s cubic-bezier(0.4, 0, 0.2, 1);
   box-sizing: border-box;
@@ -35,8 +34,8 @@
 
 .profile-stat h3 {
   font-size: 0.9rem;
-  font-weight: 600;
-  color: #525252;
+  font-weight: 700;
+  color: var(--color-primary);
   margin: 0 0 0.75rem 0;
   text-transform: uppercase;
   letter-spacing: 0.05em;
@@ -47,9 +46,9 @@
 }
 
 .profile-stat p {
-  font-size: 2.5rem;
-  font-weight: 700;
-  color: #007AFF;
+  font-size: 2rem;
+  font-weight: 500;
+  color: var(--color-text);
   margin: 0;
   line-height: 1;
 }
