+++ frontend/src/index.css
@@ -3,6 +3,7 @@
   --color-primary-light: #e3f2ff;
   --color-secondary: #8e8e93;
   --color-bg: #f2f2f7;
+  --color-warning: #f59e0b;
   --color-surface: rgba(255, 255, 255, 0.7);
   --color-text: #1c1c1e;
   --color-border: rgba(0, 0, 0, 0.08);
