+++ frontend/src/layouts/ProfileVerificationLayout/ProfileVerificationLayout.css
@@ -12,27 +12,27 @@
 .profile-verification-toggle-card {
   width: 100%;
   max-width: 100%;
-  background: rgba(255, 255, 255, 0.9);
+  background: var(--color-primary-light);
+  min-height: 120px;
   backdrop-filter: blur(20px);
   -webkit-backdrop-filter: blur(20px);
   border-radius: 20px;
-  padding: 1.5rem;
-  box-shadow: 0 4px 16px rgba(0, 0, 0, 0.08);
-  border: 2px solid rgba(0, 0, 0, 0.15);
+  padding: 1.75rem;
+  border: 1px solid var(--color-border);
   transition: all 0.3s cubic-bezier(0.4, 0, 0.2, 1);
   cursor: pointer;
   display: flex;
   align-items: center;
   justify-content: space-between;
-  gap: 1rem;
+  
   box-sizing: border-box;
   margin: 0 auto;
 }
 
 .profile-verification-toggle-card:hover {
-  transform: translateY(-2px);
-  box-shadow: 0 8px 24px rgba(0, 0, 0, 0.12);
-  border-color: rgba(0, 122, 255, 0.3);
+  transform: translateY(-4px);
+  box-shadow: 0 8px 20px rgba(0, 122, 255, 0.15);
+  border-color: var(--color-primary);
 }
 
 [data-theme="dark"] .profile-verification-toggle-card {
@@ -49,13 +49,13 @@
   flex: 1;
   display: flex;
   flex-direction: column;
-  gap: 0.35rem;
+  gap: 0.75rem;
 }
 
 .profile-verification-toggle-title {
-  font-size: 1.125rem;
-  font-weight: 600;
-  color: #1a1a1a;
+  font-size: 0.9rem;
+  font-weight: 700;
+  color: var(--color-primary);
   margin: 0;
 }
 
@@ -64,8 +64,9 @@
 }
 
 .profile-verification-toggle-subtitle {
-  font-size: 0.95rem;
-  color: #525252;
+  font-size: 0.9rem;
+  color: var(--color-text);
+  font-weight: 600;
   margin: 0;
 }
 
@@ -79,15 +80,26 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  background: rgba(0, 122, 255, 0.1);
+  background: var(--color-primary);
   border-radius: 12px;
-  color: #007AFF;
+  color: var(--color-bg);
   transition: all 0.3s cubic-bezier(0.4, 0, 0.2, 1);
 }
-
+.profile-verification-toggle-icon:hover{
+  background: var(--color-primary-light);
+  color: var(--color-primary);
+  border: 2px solid var(--color-primary);
+}
 .profile-verification-toggle-icon.expanded {
   transform: rotate(180deg);
-  background: rgba(0, 122, 255, 0.15);
+    background: var(--color-primary-light);
+  color: var(--color-primary);
+  border: 2px solid var(--color-primary);
+}
+
+.profile-verification-toggle-icon.expanded:hover{
+  background: var(--color-primary-light);
+
 }
 
 .profile-verification-items {
